PIQA-78 Risk analysis read (#97)
