+++ features/purpose/step_definitions/purpose-common-steps.ts
@@ -249,3 +249,19 @@ Given(
     }
   }
 );
+
+Then(
+  "si ottiene status code {int} e il template in versione {string}",
+  async function (statusCode: number, expectedVersion: string) {
+    assertContextSchema(this, {
+      response: z.object({
+        status: z.number(),
+        data: z.object({
+          version: z.string(),
+        }),
+      }),
+    });
+    assert.equal(this.response.status, statusCode);
+    assert.equal(this.response.data.version, expectedVersion);
+  }
+);
