+++ features/purpose/step_definitions/purpose-latest-risk-analysis-template-read.ts
@@ -1,5 +1,4 @@
-import assert from "assert";
-import { Then, When } from "@cucumber/cucumber";
+import { When } from "@cucumber/cucumber";
 import { z } from "zod";
 import {
   assertContextSchema,
@@ -15,22 +14,5 @@ When(
       await apiClient.purposes.retrieveLatestRiskAnalysisConfiguration(
         getAuthorizationHeader(this.token)
       );
-  }
-);
-
-Then(
-  "si ottiene status code {int} e il template in versione {string}",
-  async function (statusCode: number, expectedVersion: string) {
-    assertContextSchema(this, {
-      response: z.object({
-        status: z.number(),
-        data: z.object({
-          version: z.string(),
-          questions: z.array(z.unknown()),
-        }),
-      }),
-    });
-    assert.equal(this.response.status, statusCode);
-    assert.equal(this.response.data.version, expectedVersion);
   }
 );
